+++ odra-casper/wasm-env/src/lib.rs
@@ -17,7 +17,7 @@ mod wasm_contract_env;
 pub use casper_contract;
 pub use wasm_contract_env::WasmContractEnv;
 
-#[cfg(target_arch = "wasm32")]
+#[cfg(all(target_arch = "wasm32", not(feature = "disable-allocator")))]
 #[allow(unused_imports)]
 use ink_allocator;