+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Changelog for `odra`.
 ## [0.9.0] - 2024-XX-XX
 ### Added
 - `ContractRef` trait with `new` and `address` functions. All contract references now implement it.
+- `disable-allocator` feature for `odra` crate. It allows to disable the allocator used by Odra Framework in
+wasm build.
 
 ### Changed
 - Traits implemented by modules are now also implemented by their `ContractRefs` and `HostRefs`.