+++ CHANGELOG.md
@@ -2,16 +2,17 @@
 
 Changelog for `odra`.
 
-## [0.9.0] - 2024-XX-XX
+## [0.8.1] - 2024-XX-XX
 ### Added
 - `ContractRef` trait with `new` and `address` functions. All contract references now implement it.
 - `disable-allocator` feature for `odra` crate. It allows to disable the allocator used by Odra Framework in
 wasm build.
+- `odra-build` crate for including the code in contract's build.rs file.
 
 ### Changed
 - Traits implemented by modules are now also implemented by their `ContractRefs` and `HostRefs`.
 
-## [0.8.0] - 2024-XX-XX
+## [0.8.0] - 2024-02-06
 
 ### Changed
 - Replaced `contract_env::` with `self.env()` in the contract context (of type `ContractEnv`).